Move PayHandler from CompletePayFragment into MainActivity
